OB-3264 refactor
